projects/services/ingestion/web-crawl/requirements.txt
# Web Crawl Requirements
# Core crawling
crawl4ai>=0.7.0

# Data validation
pydantic>=2.0.0

# Database ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Kafka messaging
kafka-python>=2.0.0

# Environment variables
python-dotenv>=1.0.0


